Adicionando uma base postgresql no docker compose junto com um visualizador de base. Adicionando CRUD na tabela de moedas Adicionando script inicial para criacao da base Adicionando mapeamento para modelo de moeda e seu dto

+++ currency-conversion/currency-conversion.infrastructure/Data/Repositories/CurrencyRepository.cs
@@ -33,11 +33,14 @@ namespace currency_conversion.infrastructure.Data.Repositories
             return currencyUpdated;
         }
 
-        public void Delete(Currency currency)
+        public void Delete(string code)
         {
-            _dbContext.Currency.Remove(currency);
-            _dbContext.SaveChanges();
+            var currency = _dbContext.Currency.Find(code);
+            if(currency != null)
+            {
+                _dbContext.Currency.Remove(currency);
+                _dbContext.SaveChanges();
+            }
         }
-
     }
 }
